// src/lib/api/gen/types/CreateLeaveDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreateLeaveDto = {
  /**
   * @type string
   */
  leaveType: string;
  /**
   * @minLength 1
   * @type number
   */
  leaveDays: number;
  /**
   * @type string
   */
  startDate: string;
  /**
   * @type string
   */
  endDate: string;
  /**
   * @type object | undefined
   */
  reason?: object;
  /**
   * @type object | undefined
   */
  employeeGrade?: object;
  /**
   * @type object | undefined
   */
  departmentId?: object;
  /**
   * @type object | undefined
   */
  customFieldsJson?: object;
};